pyproject.toml
[tool.poetry]
name = "zivy-obraz"
version = "0.1.0"
description = "Scripts to generate screens for ePaper displays using zivyobraz.eu service."
authors = ["PavelDusek <pavel.dusek@gmail.com>"]
license = "GNU General Public License v3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.3"
lxml = "^5.1.0"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
